LayoutTests/media/modern-media-controls/tracks-panel/tracks-panel-right-x.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>TracksPanel</code> class.");

const tracksPanel = new TracksPanel;

tracksPanel.rightX = 10;
debug("tracksPanel.rightX = 10");

scheduler.frameDidFire = function()
{
    shouldBeEqualToString("tracksPanel.element.style.right", "10px");

    debug("");
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
